Create Github Action for periodical data fetching
Github Action has been created to fetch data on periodical basis. It is scheduled to run every 5 minutes and is triggered manually as well as on push of `.github/workflows/flat.yml` and `query.sql` files.

The job runs on an Ubuntu environment and installs the latest version of Deno. The repository is then checked out and data is fetched using Octo's Flat connector and saved in `data.json`. The SQL connection string is fetched from the secret repository.

Also included are new files, `.gitignore` and `README.md`.